requirements.txt
arrow==1.1.1
asgiref==3.4.1
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
fastapi==0.68.0
firebase-admin==5.0.2
fmt==0.3.1
google-api-core==1.31.2
google-api-python-client==2.17.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.2.0
google-cloud-storage==1.42.0
google-crc32c==1.1.2
google-resumable-media==2.0.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
h11==0.12.0
httplib2==0.19.1
idna==3.2
Jinja2==3.0.1
jinja2-time==0.2.0
make==0.1.6.post2
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==21.0
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
uvicorn==0.15.0
